+++ changelog.md
@@ -2,3 +2,6 @@
 * Added a script to take a full screenshot directly, bypassing the keyboard command layer. No default gesture.
 * Added a script to take a  screenshot of current window directly, bypassing the keyboard command layer. No default gesture.
 * Added  a script to take a  screenshot of focused object directly, bypassing the keyboard command layer. No default gesture.
+
+* Added script to open the folder where the screenshots are stored. No default gesture.
+* Added a script to open the file of the last screenshot with the default application. No default gesture.